Move client/web/src/auth/Steps to wildcard/components folder #25244
